steel_utils

Module codec

This module contains various codecs for reading and writing data.

Re-exports

pub use bit_set::BitSet;

pub use lp_vec3::LpVec3;

pub use or::Or;

pub use var_int::VarInt;

pub use var_long::VarLong;

pub use var_uint::VarUint;

Modules

bit_set

A module for a bit set.

glam

A module for codec impl for glam crate.

lp_vec3

A module for vanilla LpVec3 packed vector encoding.

or

A module for an Or type that can be one of two types.

var_int

A module for a variable-length integer.

var_long

A module for a variable-length long integer.

var_uint

A module for a variable-length unsigned integer.