steel_utils

Module locks

A module for custom locks. Lock wrappers for debug checks and deadlock prevention.

Structs

SteelMutex

A mutex wrapper that allows debug checks to prevent deadlocks.

SteelRwLock

An RwLock wrapper that allows debug checks to prevent deadlocks.

Traits

GenericLock

A trait for locks that can be used with the SteelLock wrapper.

Type Aliases

AsyncMutex

An asynchronous mutex.

AsyncRwLock

An asynchronous read-write lock.

SyncMutex

A synchronous mutex.

SyncRwLock

A synchronous read-write lock.