steel_utils

Module serial

This module contains traits for serializing and deserializing data.

Re-exports

pub use write::OptionalNbt;

Modules

prefixed_read

A module for reading prefixed data.

prefixed_write

A module for writing prefixed data.

read

A module for reading data.

write

A module for writing data.

Constants

DEFAULT_BOUND 🔒

Traits

PrefixedRead

A trait for reading prefixed data from a cursor.

PrefixedWrite

A trait for writing prefixed data to a writer.

ReadFrom

A trait for reading data from a cursor.

WriteTo

A trait for writing data to a writer.