steel_utils::serial

Trait PrefixedRead

pub trait PrefixedRead: Sized {
    // Required method
    fn read_prefixed_bound<P: TryInto<usize> + ReadFrom>(
        data: &mut Cursor<&[u8]>,
        bound: usize,
    ) -> Result<Self>;

    // Provided method
    fn read_prefixed<P: TryInto<usize> + ReadFrom>(
        data: &mut Cursor<&[u8]>,
    ) -> Result<Self> { ... }
}

A trait for reading prefixed data from a cursor.

Required Methods

fn read_prefixed_bound<P: TryInto<usize> + ReadFrom>( data: &mut Cursor<&[u8]>, bound: usize, ) -> Result<Self>

Reads prefixed data from a cursor with a bound.

Provided Methods

fn read_prefixed<P: TryInto<usize> + ReadFrom>( data: &mut Cursor<&[u8]>, ) -> Result<Self>

Reads prefixed data from a cursor.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".