steel_utils::serial

Trait PrefixedWrite

pub trait PrefixedWrite {
    // Required method
    fn write_prefixed_bound<P: TryFrom<usize> + WriteTo>(
        &self,
        writer: &mut impl Write,
        bound: usize,
    ) -> Result<()>;

    // Provided method
    fn write_prefixed<P: TryFrom<usize> + WriteTo>(
        &self,
        writer: &mut impl Write,
    ) -> Result<()> { ... }
}

A trait for writing prefixed data to a writer.

Required Methods

fn write_prefixed_bound<P: TryFrom<usize> + WriteTo>( &self, writer: &mut impl Write, bound: usize, ) -> Result<()>

Writes prefixed data to a writer with a bound.

Provided Methods

fn write_prefixed<P: TryFrom<usize> + WriteTo>( &self, writer: &mut impl Write, ) -> Result<()>

Writes prefixed data to a writer.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".