steel_utils::types

Struct Identifier

pub struct Identifier {
    pub namespace: Cow<'static, str>,
    pub path: Cow<'static, str>,
}

An identifier used by Minecraft.

Fields
namespace: Cow<'static, str>

The namespace of the identifier.

path: Cow<'static, str>

The path of the identifier.

Implementations

impl Identifier

pub const VANILLA_NAMESPACE: &'static str = "minecraft"

The vanilla namespace.

pub fn new( namespace: impl Into<Cow<'static, str>>, path: impl Into<Cow<'static, str>>, ) -> Self

Creates a new Identifier with the given namespace and path.

pub const fn new_static(namespace: &'static str, path: &'static str) -> Self

pub const fn vanilla(path: String) -> Self

Creates a new Identifier with the vanilla namespace.

pub const fn vanilla_static(path: &'static str) -> Self

Creates a new Identifier with the vanilla namespace and a static path.

pub const fn valid_namespace_char(char: char) -> bool

Returns whether the character is a valid namespace character.

pub const fn valid_char(char: char) -> bool

Returns whether the character is a valid path character.

pub fn validate_namespace(namespace: &str) -> bool

Returns whether the namespace is valid.

pub fn validate_path(path: &str) -> bool

Returns whether the path is valid.

pub fn validate(namespace: &str, path: &str) -> bool

Returns whether the namespace and path are valid.