Crate steel_utils

Steel Utils

This crate contains a collection of utilities used by the Steel Minecraft server.

Re-exports

pub use types::BlockPos;

pub use types::BlockStateId;

pub use types::ChunkPos;

pub use types::Identifier;

pub use types::SectionPos;

Modules

codec

This module contains various codecs for reading and writing data.

front_vec ๐

hash

CRC32C hashing for component validation. CRC32C hashing for component validation.

locks

A module for custom locks. Lock wrappers for debug checks and deadlock prevention.

math

This module contains math related utilities.

random

This module contains utilities for random number generation.

serial

This module contains traits for serializing and deserializing data.

text

This module contains everything related to text components.

translations

translations_registry

types

A module for common types.

Structs

FrontVec

Its like a vec but with reserveable front space. Its meant for our packet serialization, you can just put the len of the packet in front without reallocating keep in mind that calling multiple set_in_front() sets the data in reverse order compared to extend_from_slice()